Adds helpers for tW and tA
